requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0
extract-msg==0.47.0
reportlab==4.0.7
PyPDF2==3.0.1
python-logging-loki==1.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2